refactor: consolidate state stores into stores directory
