+++ frontend/src/components/features/chat/components/chat-input-container.tsx
@@ -3,7 +3,7 @@ import { DragOver } from "../drag-over";
 import { UploadedFiles } from "../uploaded-files";
 import { ChatInputRow } from "./chat-input-row";
 import { ChatInputActions } from "./chat-input-actions";
-import { useConversationStore } from "#/state/conversation-store";
+import { useConversationStore } from "#/stores/conversation-store";
 import { cn } from "#/utils/utils";
 
 interface ChatInputContainerProps {
