+++ frontend/src/components/features/microagent-management/microagent-management-main.tsx
@@ -1,4 +1,4 @@
-import { useMicroagentManagementStore } from "#/state/microagent-management-store";
+import { useMicroagentManagementStore } from "#/stores/microagent-management-store";
 import { MicroagentManagementDefault } from "./microagent-management-default";
 import { MicroagentManagementOpeningPr } from "./microagent-management-opening-pr";
 import { MicroagentManagementReviewPr } from "./microagent-management-review-pr";
