+++ frontend/src/components/features/microagent-management/microagent-management-view-microagent-content.tsx
@@ -1,6 +1,6 @@
 import { useTranslation } from "react-i18next";
 import { Spinner } from "@heroui/react";
-import { useMicroagentManagementStore } from "#/state/microagent-management-store";
+import { useMicroagentManagementStore } from "#/stores/microagent-management-store";
 import { useRepositoryMicroagentContent } from "#/hooks/query/use-repository-microagent-content";
 import { I18nKey } from "#/i18n/declaration";
 import { extractRepositoryInfo } from "#/utils/utils";
