+++ frontend/src/components/features/microagent-management/microagent-management-view-microagent-header.tsx
@@ -2,7 +2,7 @@ import { useTranslation } from "react-i18next";
 import { BrandButton } from "../settings/brand-button";
 import { getProviderName, constructMicroagentUrl } from "#/utils/utils";
 import { I18nKey } from "#/i18n/declaration";
-import { useMicroagentManagementStore } from "#/state/microagent-management-store";
+import { useMicroagentManagementStore } from "#/stores/microagent-management-store";
 
 export function MicroagentManagementViewMicroagentHeader() {
   const { t } = useTranslation();
