+++ frontend/src/components/features/microagent-management/microagent-management-view-microagent.tsx
@@ -1,4 +1,4 @@
-import { useMicroagentManagementStore } from "#/state/microagent-management-store";
+import { useMicroagentManagementStore } from "#/stores/microagent-management-store";
 import { MicroagentManagementViewMicroagentHeader } from "./microagent-management-view-microagent-header";
 import { MicroagentManagementViewMicroagentContent } from "./microagent-management-view-microagent-content";
