+++ frontend/src/hooks/use-auto-resize.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, RefObject, useRef } from "react";
-import { IMessageToSend } from "#/state/conversation-store";
+import { IMessageToSend } from "#/stores/conversation-store";
 import { EPS } from "#/utils/constants";
 import { getStyleHeightPx, setStyleHeightPx } from "#/utils/utils";
 import { useDragResize } from "./use-drag-resize";
