+++ frontend/src/stores/command-store.ts
